Organize settings into dedicated category screens
- Split appearance, drawer behavior, app management, and home widget settings into separate screens
- Add navigation routes and reorganize shared settings components
